Modificando tipo da chave da entidade admin de Long para Integer

+++ back-end/src/main/java/com/f1/api/domain/admin/Admin.java
@@ -6,6 +6,6 @@ import jakarta.persistence.*;
 public class Admin {
 
     @Id
-    private Long id = 1L;
+    private Integer id = 1;
 
 }
